Capture Manager with resolution and position
Working version of Camera
